# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación realizado con Swagger, servicios funcionando con Laravel 8",
        "description": "Proyecto para la comunicación REST del sistema de Inventarios",
        "contact": {
            "name": "Correo de contacto",
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/backend_inventario/public"
        }
    ],
    "paths": {
        "/service/rest/v1/catalogo/departamento": {
            "get": {
                "tags": [
                    "Departamento"
                ],
                "summary": "Muestra todos los departamentos registrados en la base de datos.",
                "description": "Retorna un array de departamentos.",
                "operationId": "getDepartaments",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/catalogo/municipio": {
            "get": {
                "tags": [
                    "Municipio"
                ],
                "summary": "Muestra todos los municipios registrados en la base de datos.",
                "description": "Retorna un array de municipios.",
                "operationId": "getMunicipalities",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Verificación de credenciales de acceso del usuario.",
                "description": "Retorna el objeto del usuario y el token.",
                "operationId": "postLogin",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Muestra la información del usuario que se encuentra en sesión.",
                "description": "Retorna un objeto del usuario en sesión.",
                "operationId": "getUserSesión",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/auth/logout": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Muestra la información del usuario que finalizará la sesión.",
                "description": "Retorna un objeto del usuario que finalizó la sesión.",
                "operationId": "getLogout",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Muestra todos los usuarios registrados en la base de datos.",
                "description": "Retorna un array de usuarios.",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Crear un nuevo usuario en el sistema.",
                "description": "Retorna el objeto del usuario creado.",
                "operationId": "postUsers",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user/{user}/edit": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar la contraseña del usuario seleccionado.",
                "description": "Retorna el objeto del usuario actualizado.",
                "operationId": "findUserbyId",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario para actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user/{user}": {
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Actualizar el usuario seleccionado.",
                "description": "Retorna el objeto del usuario actualizado.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario para actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Eliminar el usuario seleccionado.",
                "description": "Retorna el objeto del usuario eliminado.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID del usuario para eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user_rol": {
            "post": {
                "tags": [
                    "Usuario y Rol"
                ],
                "summary": "Asignar rol al usuario.",
                "description": "Retorna el objeto del rol asignado.",
                "operationId": "postUserRol",
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/service/rest/v1/security/user_rol/{user_rol} ": {
            "get": {
                "tags": [
                    "Usuario y Rol"
                ],
                "summary": "Muestra todos los roles asignados al usuario seleccionado.",
                "description": "Retorna un array de roles del usuario seleccionado.",
                "operationId": "findUserRolbyId",
                "parameters": [
                    {
                        "name": "user_rol",
                        "in": "path",
                        "description": "ID del usuario para buscar los roles",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuario y Rol"
                ],
                "summary": "Eliminar el rol asignado al usuario seleccionado.",
                "description": "Retorna el objeto del rol asignado al usuario eliminado.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user_rol",
                        "in": "path",
                        "description": "ID del rol para eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta correcta",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Permisos denegados"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "404": {
                        "description": "Servicio no encontrado"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "security": [
        []
    ]
}